@@ -2,9 +2,10 @@
   "name": "node-graphql",
   "version": "1.0.0",
   "engines": {
-    "node": "^20.0.0"
+    "node": "22.15.0"
   },
   "description": "API for yourmonthly.app",
+  "type": "module",
   "main": "start.js",
   "scripts": {
     "prisma-studio": "npx prisma studio",
@@ -15,20 +16,19 @@
   "author": "Sinisa Mikulic",
   "license": "ISC",
   "dependencies": {
-    "@prisma/client": "^4.2.1",
-    "@sentry/node": "^7.73.0",
-    "@sentry/profiling-node": "^1.2.1",
-    "apollo-server": "^3.13.0",
+    "@apollo/server": "^4.12.1",
+    "@prisma/client": "^4.16.2",
+    "@sentry/node": "^7.118.0",
+    "@sentry/profiling-node": "^1.3.5",
     "bcryptjs": "^2.4.3",
     "dotenv": "^16.3.1",
-    "esm": "^3.2.25",
-    "graphql": "^16.9.0",
+    "graphql": "^16.11.0",
     "jsonwebtoken": "^9.0.2",
     "lodash": "^4.17.21",
-    "postmark": "^3.1.1"
+    "postmark": "^4.0.5"
   },
   "devDependencies": {
-    "nodemon": "^2.0.19",
-    "prisma": "^4.2.1"
+    "nodemon": "^3.1.10",
+    "prisma": "^4.16.2"
   }
 }
