@@ -1,2 +1,5 @@
-require = require("esm")(module /*, options*/);
-module.exports = require("./src/index.js");
+// start.js
+// require = require("esm")(module /*, options*/);
+// module.exports = require("./src/index.js");
+
+import "./src/index.js";
